src/third_party/WebKit/LayoutTests/http/tests/loading/sxg/sxg-fallback-with-fragment.html

<!DOCTYPE html>
<title>Fallback redirect of SignedHTTPExchange should preserve URL fragment</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="./resources/sxg-util.js"></script>
<body>
<script>
promise_test(async (t) => {
  await waitUntilDidFinishLoadForFrame;

  // The timestamp of the test SXG file is "Apr 1 2018 00:00 UTC" and valid
  // until "Apr 8 2018 00:00 UTC". So in Apr 10, the page load should fail.
  await setSignedExchangeVerificationTime(new Date("Apr 10 2018 00:01 UTC"));

  const event = await new Promise(async (resolve, reject) => {
    // We can't catch the network error on iframe. So we use the timer.
    t.step_timeout(() => reject('timeout'), 1000);

    const frame =
        await withIframe('resources/sxg-location.sxg#fragment', 'sxg_iframe');
    const channel = new MessageChannel();
    channel.port1.onmessage = resolve;
    frame.contentWindow.postMessage(
        {port: channel.port2}, '*', [channel.port2]);
  });
  assert_equals(event.data.location, 'https://127.0.0.1:8443/loading/sxg/resources/inner-url.html#fragment');
  assert_true(event.data.is_fallback);
}, 'Fallback redirect of SignedHTTPExchange should preserve URL fragment');
</script>
</body>
